+++ internal/component/beyla/beyla_linux_test.go
@@ -1,3 +1,5 @@
+//go:build (linux && arm64) || (linux && amd64)
+
 package beyla
 
 import (
